+++ config/index.js
@@ -4,7 +4,34 @@ import config from 'services/config';
 config.set({
     app: {
         name: 'Groundwire Strategy Administration',
-        author: 'Matthew Herron'
+        author: 'Matthew Herron',
+        contributors: [
+            'Chris Hastings'
+        ],
+        pages: {
+            login: {
+                title: 'Groundwire Login',
+                headings: {
+                    main: 'Login to Groundwire Administration'
+                }
+            }
+        }
+    },
+    messages: {
+        flash: {
+            server: {
+                error: 'Something went wrong please try again'
+            },
+            login: {
+                success: 'Login successful',
+                unauthorized: 'Invalid email address or password'
+            }
+        }
+    },
+    ajax: {
+        login: {
+            url: 'https://api.groundwire.co/v1/user/login'
+        }
     }
 });