+++ src/components/login.vue
@@ -2,20 +2,18 @@
   <div class="login-container">
   	<div class="container">
   		<div class="box">
-  		  <div class="formContainer">
-  		      <h4>Connect to RobinHood Account</h4>
+  		  <div class="form-container">
+  		      <h4>{{ heading }}</h4>
   		      <div class="form-group">
-  		        <textarea class="form-control" placeholder="username" v-model="credential.username">
-  		        </textarea>
+  		        <input type="email" class="form-control" placeholder="Email Address" v-model="email" />
   		      </div>
   		      <div class="form-group">
-  		        <textarea class="form-control" placeholder="password" v-model="credential.password">
-  		       </textarea>
+  		        <input type="password" class="form-control" placeholder="Password" v-model="password" />
   		      </div>
-  		      <button class="connect" v-on:click="authenticate(credential)">
+  		      <button class="connect" v-on:click="authenticate()">
   		      Connect</button>
-			  <div class="message unloaded">
-				  {{message}}
+			  <div class="message" v-bind:class="{ error: flash.error }">
+				  {{flash.message}}
 			  </div>
   		    </div>
   		</div>
@@ -24,44 +22,73 @@
 </template>
 
 <script>
-	import model from '../state/model.js';
 	import router from '../router.js';
+	import config from 'config';
 	export default {
 	    data() {
             return {
-				credential : {
-                    username: '',
-					password: ''
-				},
-				message : '',
-				state: model
+				heading: config.get('app.pages.login.headings.main'),
+				email: null,
+				password: null,
+				flash: {
+					message: null,
+					error: false
+				}
 			}
 		},
 		methods: {
-            authenticate(arg) {
-                const credentials = model.credentials;
-                for (let i in credentials){
-                    if (arg.username === credentials[i].username && arg.password === credentials[i].password) {
-                        this.message = 'Login successful';
-                        this.state.authenticated = true;
-                        setTimeout(() => {
-                            location.href = ('/#/dashboard');
-							// router.go('/dashboard');
-						}, 2500);
-                        break;
-                    } else {
-                        this.message = 'Login failed';
+            authenticate() {
+				let formdata = new FormData();
+				formdata.append('emailAddress', this.email);
+				formdata.append('password', this.password);
+				fetch(config.get('ajax.login.url'), {
+					method: 'POST',
+					credentials: 'include',
+					mode: 'cors',
+					body: formdata
+				})
+				.then((res) => {
+					if (res.status === 401) {
+						this.flash.error = true;
+						this.flash.message = config.get('messages.flash.login.unauthorized');
+					} else if (res.status === 200) {
+						this.flash.error = false;
+						this.flash.message = config.get('messages.flash.login.success');
+					} else {
+						this.flash.error = true;
+						this.flash.message = config.get('messages.flash.server.error');
 					}
-				}
-				let target = document.getElementsByClassName('message')[0];
-                function fadeToggle(){
-                    target.className === 'message unloaded' ? target.className = 'message loaded' : target.className = 'message unloaded';
-				}
-				fadeToggle();
-				setTimeout(() => {
-                    fadeToggle();
-                }, 2000);
+					
+				})
+				.catch((err) => {
+					this.flash.error = true;
+					this.flash.message = config.get('messages.flash.server.error');
+				});
 			}
 		}
 	}
 </script>
+
+<style lang="scss" scoped>
+	@import '../assets/styles/component.scss';
+	.form-control, button.connect {
+		width: 100%;
+		height: 50px;
+		font-size: 20px;
+		margin-top: 20px;
+	}
+	button.connect {
+		background-color: $app-blue;
+		color: $app-white;
+		border: 0;
+	}
+	.form-container {
+		display: table-cell;
+		vertical-align: middle;
+	}
+	.message {
+		margin-top: 20px;
+		text-align: center;
+		font-size: 20px;
+	}
+</style>