+++ webpack.config.js
@@ -1,11 +1,12 @@
 var path = require('path');
 
 module.exports = {
-    entry: "./entry.js",
+    entry: ['whatwg-fetch', './entry.js'],
     output: {
         path: __dirname + '/dist',
         filename: "bundle.js"
     },
+    devtool: 'source-map',
     module: {
         rules: [
             {